And again some makefile stuff (this time it's some new depend code)
We are back to the behaviour of some old version of the Makefile

If you want to compile ZNC only once, just use make.

If you want the Makefile do to dependency tracking and that fancy stuff,
run make depend which creates the .depend and modules/.depend dirs.

From then on some dependency files get written when you compile something.

(The way we did it before now was bad, because all the .cpp files depended on
 the .depend dir which got its mtime updated regulary)


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@922 726aef4b-f618-498e-8847-2d620e286838