Fix compilling modperl
Modperl was broken for quite some time and no one noticed...

The modperlFLAGS include the path to some static lib which will be linked in.
For this to work, we need to already have used one of its symbols.

Now, if the linker first sees this DynaLoader.a and then our object file,
there are no references to any symbol from it yet and thus he drops it.

The fix is easy: The modperlFLAGS need to be after modperl.cpp on the
command line.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@970 726aef4b-f618-498e-8847-2d620e286838