Fix some channel key-related mess ups.
ZNC used to keep track of two channel keys: The current one from the config and
the last one it got from a +k channel mode (/mode #chan is done on join).
The +k one, if it was known, always overwrote the config one and thus the one
from the config doesnt need to be saved anymore.

This patch now makes ZNC only use one var to hold the current channel key
and this key is NOT overwritten for +k on channels.

The reason for the last change: Stupid quakenet. (RFC, anyone?)
In /mode replies quakenet always shows '*' as channel key, if one is set.
This whole patch started with someone reporting that ZNC kept forgetting the
current channel key and used '*' instead...


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1186 726aef4b-f618-498e-8847-2d620e286838