Revert rev 1204
Well... Not a complete revert, the error messages are improved for our friendly
debian users who cant find -dev packages by themselves. -.-

It turns out that on some arches (namely sparc64) libssl depends on libcrypto,
but doesn't link to it. This means that order is important when checking for
these libs. (Currently the check for libssl fails due to unresolved symbols)

Thanks to fred for reporting this.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1207 726aef4b-f618-498e-8847-2d620e286838