Handle SIGCHLD for collecting zombie processes
The old code failed to collect the following command to *shell correctly:
 exec > /dev/null 2> /dev/null < /dev/null ; sleep 10

The timeout in CExecSock::close2() was over before the child processes exited
and a zombie was left behind. Plus we used some ugly busy looping while
waiting for the child to exit. Using SIGCHLD for collecting zombies is
soooooo much nicer.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1504 726aef4b-f618-498e-8847-2d620e286838