Close client sockets only after all pending data was sent
Now e.g. the message "Invalid Password" on login even reaches the other end
before the connection shoots itself in the foot.

There is one place in Client.cpp which I didn't change: ReachedMaxBuffer()
If a client is flooding is, why should we wait until we are done answering
that flood?...

Thanks to morelia for making me notice this.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1506 726aef4b-f618-498e-8847-2d620e286838