Limit HTTP POST data size to 1MiB
We need to have an upper limit of the size of HTTP POST data. With the current
code you could just send 4 GiB of data to webadmin and ZNC would try to keep all
of this in memory.

This patch implements an upper limit for HTTP POST data of 1 MiB.

Thanks to cnu for finding this.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1559 726aef4b-f618-498e-8847-2d620e286838