Fix unloading of /bin/rm'd modules
The old code first tried to dlopen() the module again to find out if it's a
global or a user module. This could have lots of weird effects.

Now we just unload the user module and if that fails retry with the global
module (if the user got the appropriate privileges).

This removes and fixes the #warning added in the last commit.

P.S.: Oh and this is less LOCs now!


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1597 726aef4b-f618-498e-8847-2d620e286838