partyline: Send the user list on forced rejoin
When a client tries to part from a channel to which it was fixed, we sent it a
join again to make it rejoin (bad irssi for not waiting for the server's reply
to PART!).

The problem here was that we didn't send the nicklist and thus the channel
appeared empty. This fixes SF.net #2802684.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1731 726aef4b-f618-498e-8847-2d620e286838