Outsmart autoconf
When checking for libperl, evil little autoconf thought we intent to use perl
everywhere (HAHA, no!) so it went forward and added -lperl to $LIBS. Despite
this being a bad idea in general (come on, it's perl!), this also made
subsequent checks fail because the linker didn't find libperl (we reverted back
LDFLAGS to an older value after checking for perl and so the path to libperl got
lost).

Fix this by outsmarting autoconf. Let's wait for the autofail to strike back!


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@1946 726aef4b-f618-498e-8847-2d620e286838