Add aclocal.m4 and regenerate configure
The always-changing versions of the pkg-config m4 macros are mildly annoying.
To work around this, we now have a file aclocal.m4 (generated by aclocal) that
contains all the needed external macros. If you just call "autoconf", it will
use the macros from that file. "autoreconf" will call "aclocal" and "autoconf"
so that both files are up-to-date.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@2054 726aef4b-f618-498e-8847-2d620e286838