Make --with-openssl do more
It now also adds lib64 to the library search path. On 32-bit systems this won't
hurt because such a directory most likely does not exist.

Also, because we are trying to find pkg-config via openssl first, the
appropriate dirs are also added to $PKG_CONFIG_PATH.

Thanks to DarthGandalf for some good suggestions on this one.


git-svn-id: https://znc.svn.sourceforge.net/svnroot/znc/trunk@2056 726aef4b-f618-498e-8847-2d620e286838